back to a stable point after lots of playing around
couldn't get assets to work as ggez traits make it so I can't
implement

....

while writing this commit, an idea came to me.
what if ggez itself had an asset manager that
had an internal hashmap for each type of asset (as I imagine
we can't have a HashMap of type <str, ...

no, we can.
we can have a hashmap of type <str, ggez::graphics::Drawable>

i'm gonna commit what i have and try this again...
i'd hate to give each state an instance of the asset struct