Update ggez to 0.4
There were quite a few breaking changes moving from ggez 0.3 to 0.4.
This commit contains all the required changes in order to get the project
to successfully compile again.

There are still some runtime issues; these are related to the way ggez draws objects.
Previously, the centre was the origin. Now, the top-left is the origin.
I still need to go through the project and fix the positioning of all the objects.